making a generic function for removing terms
