+++ apps/modernization-ui/src/apps/search/investigation/InvestigationSearch.tsx
@@ -10,6 +10,7 @@ import { InvestigationFilterEntry } from './InvestigationFormTypes';
 import { useInvestigationSearch } from './useInvestigationSearch';
 import { InvestigationSearchResultsTable, preferences } from './result/table';
 import { ColumnPreferenceProvider } from 'design-system/table/preferences';
+import { removeTerm } from '../terms';
 
 const InvestigationSearch = () => {
     const form = useForm<InvestigationFilterEntry, Partial<InvestigationFilterEntry>>({
@@ -20,34 +21,6 @@ const InvestigationSearch = () => {
 
     const { terms } = useSearchResultDisplay();
 
-    const handleRemoveTerm = (term: Term) => {
-        const formValues = form.getValues();
-        const fieldNames = Object.keys(formValues);
-
-        const matchingField = fieldNames.find((fieldName) => fieldName === term.source);
-        if (matchingField && terms.length > 1) {
-            if (
-                matchingField === 'programAreas' ||
-                matchingField === 'jurisdictions' ||
-                matchingField === 'conditions' ||
-                matchingField === 'notificationStatuses' ||
-                matchingField === 'outbreaks' ||
-                matchingField === 'caseStatuses' ||
-                matchingField === 'processingStatuses'
-            ) {
-                form.setValue(
-                    matchingField,
-                    form.getValues()?.[matchingField]?.filter((p) => p.value !== term.value) ?? []
-                );
-            } else {
-                form.resetField(matchingField as keyof InvestigationFilterEntry);
-            }
-            search();
-        } else {
-            clear();
-        }
-    };
-
     const { options: notificationStatus } = useConceptOptions('REC_STAT', { lazy: false });
     const notificationStatusResolver = findByValue(notificationStatus);
 
@@ -55,7 +28,7 @@ const InvestigationSearch = () => {
         <ColumnPreferenceProvider id="search.investigations.preferences.columns" defaults={preferences}>
             <FormProvider {...form}>
                 <SearchLayout
-                    onRemoveTerm={handleRemoveTerm}
+                    onRemoveTerm={(term: Term) => removeTerm(form, term, search, clear, terms)}
                     criteria={() => <InvestigationSearchForm />}
                     resultsAsList={() => (
                         <SearchResultList<Investigation>
