+++ apps/modernization-ui/src/apps/search/laboratory-report/LaboratoryReportSearch.tsx
@@ -1,6 +1,6 @@
 import { FormProvider, useForm } from 'react-hook-form';
 import { SearchLayout, SearchResultList } from 'apps/search/layout';
-import { Term } from 'apps/search/terms';
+import { removeTerm, Term } from 'apps/search/terms';
 import { useSearchResultDisplay } from 'apps/search/useSearchResultDisplay';
 import { LabReport } from 'generated/graphql/schema';
 import { useLaboratoryReportSearch } from './useLaboratoryReportSearch';
@@ -22,39 +22,13 @@ const LaboratoryReportSearch = () => {
 
     const { terms } = useSearchResultDisplay();
 
-    const handleRemoveTerm = (term: Term) => {
-        const formValues = form.getValues();
-        const fieldNames = Object.keys(formValues);
-        const matchingField = fieldNames.find((fieldName) => fieldName === term.source);
-        if (matchingField && terms.length > 1) {
-            if (
-                matchingField === 'programAreas' ||
-                matchingField === 'jurisdictions' ||
-                matchingField === 'enteredBy' ||
-                matchingField === 'entryMethods' ||
-                matchingField === 'eventStatus' ||
-                matchingField === 'processingStatus'
-            ) {
-                form.setValue(
-                    matchingField,
-                    form.getValues()?.[matchingField]?.filter((p) => p.value !== term.value) ?? []
-                );
-            } else {
-                form.resetField(matchingField as keyof LabReportFilterEntry);
-            }
-            search();
-        } else {
-            clear();
-        }
-    };
-
     const { resolve: findById } = useJurisdictionOptions();
 
     return (
         <ColumnPreferenceProvider id="search.laboratory-reports.preferences.columns" defaults={preferences}>
             <FormProvider {...form}>
                 <SearchLayout
-                    onRemoveTerm={handleRemoveTerm}
+                    onRemoveTerm={(term: Term) => removeTerm(form, term, search, clear, terms)}
                     criteria={() => <LaboratoryReportSearchCriteria />}
                     resultsAsList={() => (
                         <SearchResultList<LabReport>
