+++ apps/modernization-ui/src/apps/search/terms/index.ts
@@ -2,3 +2,4 @@ export * from './terms';
 export * from './fromValue';
 export * from './fromSelectable';
 export * from './fromSelectables';
+export * from './removeTerm';
